+++ R/zzz-autoroxy.R
@@ -1,3 +1,4 @@
+# nolint start
 {
     if (!"DESCRIPTION" %in% dir()) {
         return(invisible(NULL))
@@ -15,5 +16,12 @@
             call. = FALSE)
         return(invisible(NULL))
     }
-    roxygen2::roxygenize(roclets = c("rd"))
+    if (!requireNamespace("devtools")) {
+        warning("Cannot load devtools. Package documentation will be unavailable.",
+            call. = FALSE)
+        return(invisible(NULL))
+    }
+    dir.create("man")
+    devtools::document(roclets = c("rd"))
 }
+# nolint end