+++ tests/testthat/env/src/B.R
@@ -1,0 +1,4 @@
+# Single dependency
+darn::init("A", env_vars = "TWENTYONE")
+
+darn::done(twentyone = TWENTYONE)