+++ tests/testthat/subdir/dir/B.R
@@ -1,4 +1,4 @@
 # Single dependency
-path_info <- darn::init("A")
+darn::init("A")
 
 darn::done(twentyone = fortytwo / 2)